Moved classes from iToolkitService.php to their own files for PSR-4 compliance.
