+++ ToolkitApi/autoload.php
@@ -7,16 +7,16 @@ spl_autoload_register(function($class){
         'ToolkitApi\Toolkit'            => __DIR__ . DIRECTORY_SEPARATOR . 'Toolkit.php',
         'ToolkitApi\db2supp'            => __DIR__ . DIRECTORY_SEPARATOR . 'Db2supp.php',
         'ToolkitApi\httpsupp'           => __DIR__ . DIRECTORY_SEPARATOR . 'httpsupp.php',
-        'ToolkitApi\DateTimeApi'        => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\ListFromApi'        => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\UserSpace'          => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\TmpUserSpace'       => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\DataQueue'          => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\SpooledFiles'       => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\JobLogs'            => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\ObjectLists'        => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\SystemValues'       => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
-        'ToolkitApi\DataArea'           => __DIR__ . DIRECTORY_SEPARATOR . 'iToolkitService.php',
+        'ToolkitApi\DateTimeApi'        => __DIR__ . DIRECTORY_SEPARATOR . 'DateTimeApi.php',
+        'ToolkitApi\ListFromApi'        => __DIR__ . DIRECTORY_SEPARATOR . 'ListFromApi.php',
+        'ToolkitApi\UserSpace'          => __DIR__ . DIRECTORY_SEPARATOR . 'UserSpace.php',
+        'ToolkitApi\TmpUserSpace'       => __DIR__ . DIRECTORY_SEPARATOR . 'TmpUserSpace.php',
+        'ToolkitApi\DataQueue'          => __DIR__ . DIRECTORY_SEPARATOR . 'DataQueue.php',
+        'ToolkitApi\SpooledFiles'       => __DIR__ . DIRECTORY_SEPARATOR . 'SpooledFiles.php',
+        'ToolkitApi\JobLogs'            => __DIR__ . DIRECTORY_SEPARATOR . 'JobLogs.php',
+        'ToolkitApi\ObjectLists'        => __DIR__ . DIRECTORY_SEPARATOR . 'ObjectLists.php',
+        'ToolkitApi\SystemValues'       => __DIR__ . DIRECTORY_SEPARATOR . 'SystemValues.php',
+        'ToolkitApi\DataArea'           => __DIR__ . DIRECTORY_SEPARATOR . 'DataArea.php',
         'ToolkitApi\odbcsupp'           => __DIR__ . DIRECTORY_SEPARATOR . 'Odbcsupp.php',
         'ToolkitApi\ToolkitPcml'        => __DIR__ . DIRECTORY_SEPARATOR . 'ToolkitPCML.php',
         'ToolkitApi\ProgramParameter'   => __DIR__ . DIRECTORY_SEPARATOR . 'ToolkitServiceParameter.php',
